refactor(harness): extract /api/auth routes from server.create_app
Move the session/RBAC Users-panel surface into harness/auth_routes.py using
the same register-and-inject pattern as gate_ops.py and agent_routes.py.
create_app keeps the limiter, CSRF token, and dependency lists; route
bodies no longer live in the factory. Fail-closed 503 when auth is off,
loopback-only bootstrap, and no APIRouter. Behavior unchanged.
